refactor(sundial): host Context Sync as a StreamTabView tab in the library
Move the SundialKitStreamContext "Context Sync" demo out of the
SundialStream app target and into the SundialDemoStream library, split
one type per file and sorted by role (Models/ViewModels/Views/App).

- StreamTabView keeps its tabs and now leads with a Context Sync tab
  (StreamContextView), gated for iOS 18 / watchOS 11; the old sheet hack
  is gone and SundialStreamAppMain is a thin @main rendering StreamTabView.
- Rename StreamRootView -> StreamContextView (App/).
- Add SundialKitCore/SundialKitConnectivity/SundialKitStreamContext deps to
  the SundialDemoStream target; guard the demo with #if os(iOS) || os(watchOS)
  to keep the macOS library build clean.
- Trim the SundialStream app targets to depend only on SundialDemoStream;
  drop the now-redundant project-level SundialKit/SundialKitStream packages
  (resolved transitively via the local Sundial package's branch pin).
